rem. logging, add embed IOS.
replaced logging with new ios_log that takes over console.

+++ tools/app5.js
@@ -167,19 +167,33 @@ if (args[0]=="ios") {
 		FileUtils.copyDirectory(new File(APP5_HOME+"/lib"),targetwww);
 		FileUtils.copyDirectory(new File("./www"),targetwww);
 	
-		// copy ios_wrapper to target.
-		var ios_wrapper=new File(APP5_HOME+"/ios_wrapper");
-		var target=new File('./target/ios_wrapper');
-		FileUtils.copyDirectory(ios_wrapper,target);
+		var ios_wrapper_target=new File("./target/ios_wrapper");
+		if (ios_wrapper_target.exists()) {
+			var targetwww=new File("./target/ios_wrapper/www");
+			FileUtils.copyDirectory(new File("./target/www"),targetwww);
+		}
+		else {
 		
-		// fill ios_wrapper www
-		var targetwww=new File("./target/ios_wrapper/www");
-		FileUtils.copyDirectory(new File("./target/www"),targetwww);
+			// copy ios_wrapper to target.
+			var ios_wrapper=new File(APP5_HOME+"/ios_wrapper");
+			var target=new File('./target/ios_wrapper');
+			FileUtils.copyDirectory(ios_wrapper,target);
+			
+			// fill ios_wrapper www
+			var targetwww=new File("./target/ios_wrapper/www");
+			FileUtils.copyDirectory(new File("./target/www"),targetwww);
+			
+			// rename xcode project.
+			var projectName=new File((new File('.')).getCanonicalPath()).getName();
+			var project=new File('./target/ios_wrapper');
+			renameXCodeProject(project,'ios_wrapper',projectName);
+		}
+
+	   	var s=""+FileUtils.readFileToString(new File('./target/ios_wrapper/www/index.html'));
+	   	s=s.replace(/<!--IOS EMBED -->/g,'<script>App5.require("ios_embed")</script>');
+	   	FileUtils.write(new File('./target/ios_wrapper/www/index.html'),s);
 		
-		// rename xcode project.
-		var projectName=new File((new File('.')).getCanonicalPath()).getName();
-		var project=new File('./target/ios_wrapper');
-		renameXCodeProject(project,'ios_wrapper',projectName);
+		
 	}
 	
 	quit();
